fix tags not reseting
